Changed project structure to be split up in Multi, Mac OS X & iOS classes

+++ ABFramework/Multi/Categories/NSArray+ABFramework.m
@@ -22,7 +22,7 @@
     {
         return [self objectAtIndex:index];
     }
-    NSLog(@"NSArray+ABFramework: Warning -> Attempted to access on existent index[%i]", index);
+    NSLog(@"NSArray+ABFramework: Warning -> Attempted to access on existent index[%li]", (long)index);
     return nil;
 }
